requirements.txt
asgiref==3.4.1
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
decorator==5.0.9
dnspython==2.1.0
fastapi==0.68.0
h11==0.12.0
mongoengine==0.23.1
passlib==1.7.4
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.12.0
python-dotenv==0.19.0
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
validators==0.18.2
